Make devstack work correctly behing a proxy
Devstack fails to stack properly when running in a
network with a non-transparent http proxy. This is due to the
"http_proxy" enviroment variable not being preserved when running sudo
in the devstack plugin. Add the -E flag to sudo whenever appropriate and
remove sudo when unnecessary in order to fix the issue.

Change-Id: I516076e78434ee241292b1fed6a6edcf0235d4b1
Signed-off-by: Marco Chiappero <marco.chiappero@intel.com>